corpus/benign/mcp/gh-sircharan-zerodha-mcp-trading/requirements.txt
kiteconnect==4.2.0
pandas==2.1.0
numpy==1.24.3
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
websocket-client==1.6.1
plotly==5.16.1
dash==2.13.0
openai==1.3.0
python-dateutil==2.8.2
pytz==2023.3
schedule==1.2.0
loguru==0.7.0
pytest==7.4.2
black==23.7.0
flake8==6.1.0
isort==5.12.0 